requirements.txt
Django
Prettyconf
dj-database-url
django-model-utils
django-cities-light
psycopg2-binary
django-admin-interface
gunicorn
psycogreen
gevent
whitenoise
django-extended-choices
django-anymail[mailgun]
django-templated-email
pagseguro